Fixed callsigns with tokens not being able to connect to the lobby.
Callsigns with tokens previously (R5 RC3) could not connect to the lobby. The lobby took the callsign, stripped off the token and then asked ASGS for confirmation that the authentication ticket fits the callsign. ASGS however only knew about an authentication of the callsign including the token and therefore refused.

The token is now no longer stripped off, but rather is the full callsign being urlencoded (just as the ticket already was, using the function Radar has written).

Note: This has been pushed out to the main lobby on 2009-11-18 as a fix after the official R5 deployment.

git-svn-id: file:///media/data/aa/dev/Allegiance-SVN/branch/FAZR5Contrib@501 8c327b6c-a73e-8745-8738-111f1e176ae1