Added a workaround for a problem that caused the hit points and the armor class to be no longer shown at the loadout screen.
These bars were actually always rendered, but cut off because the pane was too small. The pane has been enlarged now. Which change in R5 caused this is unknown at this point.

Additionally, the positioning of the bars of rows 2 and 4 and the text of rows 1 and 4 has been improved.

This change should both be pushed out with the next release and included in the installer.

git-svn-id: file:///media/data/aa/dev/Allegiance-SVN/branch/FAZR6@507 8c327b6c-a73e-8745-8738-111f1e176ae1