Patch to prevent 8960551
Looks like we are seeing this crash a lot (count 119.) We are getting here with an empty Device that cant get to the Device part of the next call. Please take a look at this as my dev PC is currently dead in the water. Make sure it doesn't harm the normal operation of effects (like the fire coming out of the back while thrusting)